--- Eat&Beat/Forms/User.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Eat_Beat.Forms
-{
-    public class User
-    {
-        public int idUser { get; set; }
-        public string name { get; set; }
-        public string email { get; set; }
-        public string password { get; set; }
-        public double rating { get; set; }
-    }
-}
